mfi_ddb_package/requirements.txt
cmake>=4.0.3
numpy>=1.24.4
omegaconf>=2.3.0
open3d>=0.19.0
opencv-python>=4.11.0.86
paho-mqtt>=2.1.0
pandas>=2.0.3
ping3>=4.0.8
protobuf>=5.29.5
pyyaml>=6.0.2
requests>=2.32.4
rospkg>=1.6.0
setuptools>=75.3.2
watchdog>=4.0.2
wheel>=0.45.1
xmltodict>=0.14.2
pydantic>=2.10.6
